requirements.txt
Flask==3.0.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
distro==1.9.0
github_issue_analyzer @ git+https://github.com/RaschidJFR/github-issue-analyzer@af41631f9b8b524e1222cdd335b6d479323f118e
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
numpy==2.3.1
openai==1.93.0
pandas==2.3.0
pydantic==2.11.7
pydantic_core==2.33.2
PyDispatcher==2.0.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.0.1
